Marker rating new

+++ app/src/main/java/com/mco/accessability/models/CombinedReviewModel.kt
@@ -2,7 +2,7 @@ package com.mco.accessability.models
 
 data class CombinedReviewModel(
     val placeName: String = "",
-    val placeRating: Int = 0,
+    val placeRating: Float = 0f,
     val reviewNotes: String = "",
     val author: String = ""
 )
